Restructure the packages a bit and use Tung's github ID
