+++ server/go.mod
@@ -1,10 +1,9 @@
-module github.com/luxas/favorgiver
+module github.com/thechosenoneneo/favor-giver
 
 go 1.14
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
-	github.com/go-chi/chi v4.1.1+incompatible // indirect
 	github.com/jinzhu/gorm v1.9.12
 	github.com/labstack/echo v3.3.10+incompatible
 	github.com/labstack/gommon v0.3.0 // indirect
